+++ src/main/gui/Hand.java
@@ -10,7 +10,7 @@ import processing.core.PImage;
 import processing.core.PVector;
 
 import static main.Main.*;
-import static main.util.MiscMethods.roundTo;
+import static main.util.MiscMethods.*;
 
 public class Hand {
 
@@ -277,7 +277,7 @@ public class Hand {
         }
         money -= price;
         if (changeHeld) held = "null";
-//        path.nodeCheckObs();
+        updateNodes();
         updateTowerArray();
     }
 }
