+++ src/main/pathfinding/PathRequest.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 
 import static main.Main.nSize;
 import static main.Main.nodeGrid;
-import static main.pathfinding.UpdateNodes.updateNodes;
+import static main.pathfinding.UpdateNode.updateNode;
 import static processing.core.PApplet.round;
 
 public class PathRequest {
@@ -33,7 +33,7 @@ public class PathRequest {
         //issue here?
         nodeGrid[round((enemy.position.x-(enemy.size.x/2f))/nSize)][round((enemy.position.y-(enemy.size.y/2f))/nSize)].setStart(round((enemy.position.x-(enemy.size.x/2f))/nSize), round((enemy.position.y-(enemy.size.y/2f))/nSize));
         Main.start.findGHF();
-        updateNodes(Main.start,this);
+        updateNode(Main.start,this);
         Main.path.done = false;
         Main.path.find(id); //points are added here
         enemy.swapPoints(false);
