+++ src/main/towers/turrets/EnergyBlaster.java
@@ -8,6 +8,7 @@ import processing.core.PImage;
 import processing.core.PVector;
 
 import static main.Main.*;
+import static main.util.MiscMethods.updateTowerArray;
 import static processing.core.PConstants.HALF_PI;
 
 public class EnergyBlaster extends Turret{
