+++ src/main/towers/turrets/MagicMissileer.java
@@ -8,6 +8,7 @@ import processing.core.PImage;
 import processing.core.PVector;
 
 import static main.Main.*;
+import static main.util.MiscMethods.updateTowerArray;
 
 public class MagicMissileer extends Turret{
