+++ src/main/towers/turrets/Slingshot.java
@@ -7,6 +7,7 @@ import processing.core.PImage;
 import processing.core.PVector;
 
 import static main.Main.*;
+import static main.util.MiscMethods.updateTowerArray;
 
 public class Slingshot extends Turret{
